+++ src/ExaPF.jl
@@ -24,7 +24,7 @@ using TimerOutputs
 export solve
 
 # Import submodules
-include("helpers.jl")
+include("utils.jl")
 include("ad.jl")
 using .AD
 include("algorithms/precondition.jl")
@@ -33,11 +33,11 @@ include("indexes.jl")
 using .IndexSet
 include("iterative.jl")
 using .Iterative
-include("parse/parse_mat.jl")
+include("parsers/parse_mat.jl")
 using .ParseMAT
-include("parse/parse_psse.jl")
+include("parsers/parse_psse.jl")
 using .ParsePSSE
-include("powersystem.jl")
+include("PowerSystem/PowerSystem.jl")
 using .PowerSystem
 
 const TIMER = TimerOutput()
@@ -45,13 +45,6 @@ const TIMER = TimerOutput()
 const VERBOSE_LEVEL_HIGH = 1
 const VERBOSE_LEVEL_MEDIUM = 2
 const VERBOSE_LEVEL_LOW = 3
-
-struct ConvergenceStatus
-    has_converged::Bool
-    n_iterations::Int
-    norm_residuals::Float64
-    n_linear_solves::Int
-end
 
 
 """
@@ -67,79 +60,6 @@ function residualFunction(V, Ybus, Sbus, pv, pq)
          real(mis[pq])
          imag(mis[pq]) ]
     return F
-end
-
-function residualFunction_real!(F, v_re, v_im,
-                                ybus_re, ybus_im, pinj, qinj, pv, pq, nbus)
-
-    npv = size(pv, 1)
-    npq = size(pq, 1)
-
-    # REAL PV
-    for i in 1:npv
-        fr = pv[i]
-        F[i] -= pinj[fr]
-        for (j,c) in enumerate(ybus_re.colptr[fr]:ybus_re.colptr[fr+1]-1)
-            to = ybus_re.rowval[c]
-            F[i] += (v_re[fr]*(v_re[to]*ybus_re.nzval[c] - v_im[to]*ybus_im.nzval[c]) +
-                     v_im[fr]*(v_im[to]*ybus_re.nzval[c] + v_re[to]*ybus_im.nzval[c]))
-        end
-    end
-
-    # REAL PQ
-    for i in 1:npq
-        fr = pq[i]
-        F[npv + i] -= pinj[fr]
-        for (j,c) in enumerate(ybus_re.colptr[fr]:ybus_re.colptr[fr+1]-1)
-            to = ybus_re.rowval[c]
-            F[npv + i] += (v_re[fr]*(v_re[to]*ybus_re.nzval[c] - v_im[to]*ybus_im.nzval[c]) +
-                           v_im[fr]*(v_im[to]*ybus_re.nzval[c] + v_re[to]*ybus_im.nzval[c]))
-        end
-    end
-
-    # IMAG PQ
-    for i in 1:npq
-        fr = pq[i]
-        F[npv + npq + i] -= qinj[fr]
-        for (j,c) in enumerate(ybus_re.colptr[fr]:ybus_re.colptr[fr+1]-1)
-            to = ybus_re.rowval[c]
-            F[npv + npq + i] += (v_im[fr]*(v_re[to]*ybus_re.nzval[c] - v_im[to]*ybus_im.nzval[c]) -
-                                 v_re[fr]*(v_im[to]*ybus_re.nzval[c] + v_re[to]*ybus_im.nzval[c]))
-        end
-    end
-
-    return F
-end
-function residualFunction_polar_sparsity!(F, v_m, v_a,
-                     ybus_re, ybus_im,
-                     pinj, qinj, pv, pq, nbus)
-
-    npv = size(pv, 1)
-    npq = size(pq, 1)
-
-    for i in 1:length(pv)+length(pq)
-        # REAL PQ: (npv+1:npv+npq)
-        # IMAG PQ: (npv+npq+1:npv+2npq)
-        fr = (i <= npv) ? pv[i] : pq[i - npv]
-        F[i] -= pinj[fr]
-        if i > npv
-            F[i + npq] -= qinj[fr]
-        end
-        for c in ybus_re.colptr[fr]:ybus_re.colptr[fr+1]-1
-            to = ybus_re.rowval[c]
-            aij = v_a[fr] - v_a[to]
-            # f_re = a * cos + b * sin
-            # f_im = a * sin - b * cos
-            coef_cos = v_m[fr]*v_m[to]*ybus_re.nzval[c]
-            coef_sin = v_m[fr]*v_m[to]*ybus_im.nzval[c]
-            cos_val = cos(aij)
-            sin_val = sin(aij)
-            F[i] += coef_cos * cos_val + coef_sin * sin_val
-            if i > npv
-                F[npq + i] += coef_cos * sin_val - coef_sin * cos_val
-            end
-        end
-    end
 end
 
 @kernel function residual_kernel!(F, v_m, v_a,
@@ -236,136 +156,6 @@ function project_constraints!(u::AbstractArray, grad::AbstractArray, u_min::Abst
             grad[i] = 0.0
         end
     end
-end
-
-"""
-    get_bound_constraints(pf)
-
-Given PowerNetwork object, returns vectors xmin, xmax, umin, umax
-of the OPF box constraints.
-
-"""
-function get_bound_constraints(pf::PowerSystem.PowerNetwork)
-
-    BUS_I, BUS_TYPE, PD, QD, GS, BS, BUS_AREA, VM, VA, BASE_KV, ZONE, VMAX, VMIN,
-    LAM_P, LAM_Q, MU_VMAX, MU_VMIN = IndexSet.idx_bus()
-    GEN_BUS, PG, QG, QMAX, QMIN, VG, MBASE, GEN_STATUS, PMAX, PMIN, PC1, PC2, QC1MIN,
-    QC2MIN, QC2MAX, RAMP_AGC, RAMP_10, RAMP_30, RAMP_Q, APF, MU_PMAG, MU_PMIN, MU_QMAX,
-    MU_QMIN = IndexSet.idx_gen()
-
-    nref = length(pf.ref)
-    npv = length(pf.pv)
-    npq = length(pf.pq)
-    b2i = pf.bus_to_indexes
-
-    gens = pf.data["gen"]
-    baseMVA = pf.data["baseMVA"][1]
-    bus = pf.data["bus"]
-    ngens = size(gens)[1]
-
-    dimension_u = 2*npv + nref
-    dimension_x = 2*npq + npv
-
-    u_min = fill(-Inf, dimension_u)
-    u_max = fill(Inf, dimension_u)
-    x_min = fill(-Inf, dimension_x)
-    x_max = fill(Inf, dimension_x)
-    p_min = fill(-Inf, nref)
-    p_max = fill(Inf, nref)
-
-    for i in 1:length(pf.pq)
-        bus_idx = pf.pq[i]
-        vm_max = bus[bus_idx, VMAX]
-        vm_min = bus[bus_idx, VMIN]
-        x_min[i] = vm_min
-        x_max[i] = vm_max
-    end
-
-    for i in 1:length(pf.pv)
-        bus_idx = pf.pv[i]
-        vm_max = bus[bus_idx, VMAX]
-        vm_min = bus[bus_idx, VMIN]
-        u_min[nref + npv + i] = vm_min
-        u_max[nref + npv + i] = vm_max
-    end
-
-    for i in 1:length(pf.ref)
-        bus_idx = pf.ref[i]
-        vm_max = bus[bus_idx, VMAX]
-        vm_min = bus[bus_idx, VMIN]
-        u_min[i] = vm_min
-        u_max[i] = vm_max
-    end
-
-    for i = 1:ngens
-        genbus = b2i[gens[i, GEN_BUS]]
-        bustype = bus[genbus, BUS_TYPE]
-        if bustype == PowerSystem.PV_BUS_TYPE
-            idx_pv = findfirst(pf.pv.==genbus)
-            u_min[nref + idx_pv] = gens[i, PMIN] / baseMVA
-            u_max[nref + idx_pv] = gens[i, PMAX] / baseMVA
-        elseif bustype == PowerSystem.REF_BUS_TYPE
-            idx = findfirst(pf.ref .== genbus)
-            p_min[idx] = gens[i, PMIN] / baseMVA
-            p_max[idx] = gens[i, PMAX] / baseMVA
-        end
-    end
-
-    return u_min, u_max, x_min, x_max, p_min, p_max
-end
-
-"""
-    get_bound_reactive_power(pf)
-
-Given PowerNetwork object, return bounds on reactive power
-of the PV buses.
-
-Matpower specifies bounds on the reactive power of the generator:
-
-    Q_min <= Qᵍ <= Q_max
-
-ExaPF uses internally the reactive power at the buses, which for PV buses
-writes out
-
-    Q = Qᵍ - Qˡ
-
-with Qˡ the (constant) reactive load.
-This function corrects the bounds to take into account the reactive load:
-
-    Q_min - Qˡ <= Q <= Q_max - Qˡ
-
-"""
-function get_bound_reactive_power(pf::PowerSystem.PowerNetwork)
-    BUS_I, BUS_TYPE, PD, QD, GS, BS, BUS_AREA, VM, VA, BASE_KV, ZONE, VMAX, VMIN,
-    LAM_P, LAM_Q, MU_VMAX, MU_VMIN = IndexSet.idx_bus()
-    GEN_BUS, PG, QG, QMAX, QMIN, VG, MBASE, GEN_STATUS, PMAX, PMIN, PC1, PC2, QC1MIN,
-    QC2MIN, QC2MAX, RAMP_AGC, RAMP_10, RAMP_30, RAMP_Q, APF, MU_PMAG, MU_PMIN, MU_QMAX,
-    MU_QMIN = IndexSet.idx_gen()
-
-    nref = length(pf.ref)
-    npv = length(pf.pv)
-    npq = length(pf.pq)
-
-    gens = pf.data["gen"]
-    baseMVA = pf.data["baseMVA"][1]
-    bus = pf.data["bus"]
-    ngens = size(gens)[1]
-    # Reactive load
-    qload = imag.(pf.sload)
-
-    q_min = fill(-Inf, npv)
-    q_max = fill(Inf, npv)
-
-    for i = 1:ngens
-        genbus = pf.bus_to_indexes[gens[i, GEN_BUS]]
-        bustype = bus[genbus, BUS_TYPE]
-        if bustype == PowerSystem.PV_BUS_TYPE
-            idx_pv = findfirst(pf.pv.==genbus)
-            q_min[idx_pv] = gens[i, QMIN] / baseMVA - qload[idx_pv]
-            q_max[idx_pv] = gens[i, QMAX] / baseMVA - qload[idx_pv]
-        end
-    end
-    return q_min, q_max
 end
 
 function cost_function(pf::PowerSystem.PowerNetwork, x::AbstractArray, u::AbstractArray,
@@ -620,9 +410,9 @@ function solve(
                             ybus_re, ybus_im,
                             pbus, qbus, pv, pq, nbus)
     # Build the AD Jacobian structure
-    stateJacobianAD = AD.StateJacobianAD(residualFunction_polar_sparsity!, F, Vm, Va,
+    stateJacobianAD = AD.StateJacobianAD(F, Vm, Va,
                                          ybus_re, ybus_im, pbus, qbus, pv, pq, ref, nbus)
-    designJacobianAD = AD.DesignJacobianAD(residualFunction_polar_sparsity!, F, Vm, Va,
+    designJacobianAD = AD.DesignJacobianAD(F, Vm, Va,
                                            ybus_re, ybus_im, pbus, qbus, pv, pq, ref, nbus)
     if verbose_level >= VERBOSE_LEVEL_MEDIUM
         print("State Jacobian  --- ")
