+++ src/PowerSystem/PowerSystem.jl
@@ -1,0 +1,19 @@
+module PowerSystem
+
+using Printf
+using SparseArrays
+using ..ExaPF: ParsePSSE, ParseMAT, IndexSet, Spmat
+
+import Base: show
+
+const PQ_BUS_TYPE = 1
+const PV_BUS_TYPE = 2
+const REF_BUS_TYPE  = 3
+
+
+abstract type AbstractPowerSystem end
+
+include("network_topology.jl")
+include("power_network.jl")
+
+end
